Addpmft Float, removal of excess cost comment box

+++ tool/forms.py
@@ -114,7 +114,7 @@ class PMFTForm(forms.ModelForm):
     pmftComment = forms.CharField(max_length=100,  required=False)
     pmftCost = forms.FloatField(required=False)
     pmftCostComment =  forms.MultipleChoiceField(choices=pmftCostCommentOptions)
-    pmftQuantity = forms.IntegerField()
+    pmftQuantity = forms.FloatField()
     pmftType = forms.MultipleChoiceField(choices=pmftTypeOptions)
     pmftSlug = forms.SlugField(widget=forms.HiddenInput(), required=False)
     class Meta:
